phase3-backend/pyproject.toml
[project]
name = "phase3-backend"
version = "1.0.0"
description = "Phase 3 Chat Backend - AI-Powered Todo Chatbot"
requires-python = ">=3.13"
readme = "README.md"

dependencies = [
    "fastapi>=0.122.0",
    "uvicorn[standard]>=0.27.0",
    "sqlmodel>=0.0.24",
    "asyncpg>=0.29.0",
    "openai>=2.11.0",
    "openai-agents>=0.6.0",
    "httpx>=0.27.0",
    "pyjwt[crypto]>=2.8.0",
    "cryptography>=41.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.27.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.13"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
